PartsOpt: Move PACing of globals function to ctors
Instead of inserting the function that pac the globals at the
beginning of main, declare it a constructor function with
highest priority possible 0.
These also resolve the problem on when to pac global pointers
in DSOs (e.g: shared libraries)
Notice also that user constructors can only have priority more or
equal than 100. This ensures us that we can pac any pointer before
the user constructors are called.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>