PartsOpt:test: Support global pacing constructor
Now that we do not insert the pacing function at the
beginning of main, we need to fix the tests to take
this into account.
Also add in the old test a check for inserting the pacing
function in the global constructors array.
For now, let's check that we do not insert back the pacing
function at the beginning of main. In the future, that check
can be removed.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>