PartsOpt:GlobalPass: __patuh_pac_globals() private
Declare __pauth_pac_globals private linkage. This means
that the function is only for the current Module and no
symbols should be emitted for this function.

We can revised the decision for the symbols later on, an
mark it as internal, if we need to export the symbol for
some reason.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>