PartsCpiPass:test: Fix tailcall test
Fix tailcall test. We currently harcoded some
of the physical registers expected in the instructions.
At least, I think we can rely that the use of x0 and x19
will not change. For x8, I am currently not so sure.
X0 has in the prologue of the function the call pointer.
X19 is used to avoid the clobber of X0 in the second test.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>