AArch64Parts: use fixed reg for auth modifier
This change always uses a reserved register to store the modifier when
converting indirect pointer calls. In order to accommodate register
allocations this occasionally results in two extra movs, one to move the
modifier into x23, and another to move the pointer register to another
(safe) one. The extra movs can (and should?) be optimized away in a
separate pass.