PartsCpiPass: Use register Scavenger for modifier
Finally we are approaching our final goal :-)
All those problems that we have, are due to not
be able to state cleary the dependencies between our
intrinsic and the branch instruction. Specially problematic
was the modifier value because introduces a second register
not used in non authenticated branch intrstructions.

Instead of relying on the register used to pass the modifer
value to the instrinsic, let's allocated a new register with
the help of the register scavenger for the modifier.

The current implementation just moves the modifier argument
into new allocated register. The authencitcated branch
instruction will use this new allocated register for the
modifier. Of course, this is just the initial implementation
and there is a lot of room for improvement, like removing
the mov.

We reserve in the stack an emergency spill slot in case there
are no free registers to use and we need to push some register
to the stack.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>