PartsCpiPass: Use available reg, remove spill slot
We were trying to spill registers when there were plenty
available. This was due the register scavenger backwards
iterator was located in the call instruction instead of
instruction before. Call instruction will mark a huge amount
of registers are being alive, forcing us in the case that
we use all the scratch registers to spill a callee saved
register.

Also the logic for reserving a spill slot in the stack is wrong.
It is illegal to allocate a spill slot after the Prolog Epilog
Inserter (PEI) pass has been run. This was the reason for the behaviour
where no spill slot was reserved in the stack.
So let's remove the support for spilling registers. If we do need
to do that we will fail compilation for now.
Later on we can try to add a pass before the PEI to allocate the
emergency spill slot for us. Alternatively we can try to manage
the spilling ourselves in this pass witht out relying on the register
Scavenger.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>