PartsCpiPass:test: Fix some hardcoded values
Fix some hardcoded typeid values.Use a regexp instead
cause those values may change during development.

Also add a new test for indirect tailcall that no returns
which expects one argument. We want to check that
x0 is not used for the indirect tailcall, cause it
is going to be used for passing the argument to the
tail call.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>