PartsCpiPass: Extract function for compilation error
Extract a function to handle the case we find an orphan
autcall intrinsic, in other words, without a matching
blr or br instruction.

I wonder if this can happen in a legal case were the compiler
is capable to deduce the address of the indirect call on compilation
time.

Let's leave now for development as we want to check such cases.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>