AArch64EarlyPartsCpiPass: Skeleton for Early CPI
Add boilerplate for a function pass related to PARTS.
It will be meant to be run during ILP (Instruction Level
Parallelism) optimizations.

The idea is that at this point of the compilation pipeline
PAUTH_AUTCALL and the BLR/BR associated branch to it will still
be in the basic block. This makes our life easy to transform
the BLR to a BLRA[AB].
It also help us in the case of register usage, cause at this
point we should have not done the register allocation.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>