AArch64PartsEarlyCpi:test: Fix broken test
The test was wrongly accepting the case that
we were using x19 as destination register for
the indirect branch.
In fact the commit that introduce the error should
only have disable the redundant mov check which was
the only expected outcome of handling tailcalls
before register allocation has been done.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>