AArch64PartsDpiPass: don't ignore no-parts functions
This attribute is a bit unfortunately named, the purpose is to prevent
PARTS from instrumenting functions such as the globals fixer, as it relies
on special behavior not compatible with regular PARTS stuff (i.e., it needs
to load pointers without authenticating them). Nonetheless, the backend
should still handle these files properly.