PartsSpillPass: Add data pointer spill test
Add a test that spills a data pointer into the stack.
For some reason if we have more than 8 arguments in the function
signature, LLVM will only try to use one of the callee saved registers
instead of 2 to keep the pointer arguments in registers.
This needs to be understand why these is happening, probably we can
not rely on this behaviour.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>