DataPointerArgs: Boilerplate to mark data pointers
Add skeleton for the new Parts DataPointerArgsPass.
We will use this pass to mark all the data pointers arguments
of a function.
This will be later use when lowering the function to know
if a given data pointer is temporarily spill into the stack
or not. Data pointer spills need to be protected by Parts.

NOTE: Our pass has a dependency to DeadArgumentEliminationPass.
That pass remove dead arguments of inner functions and needs
to be run before our pass. Otherwise, if we mark with an intrinsic
the data pointers the DAE may conclude that the argument is not
dead although, it is,

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>