DataPointerArgs: Extract a function
Extract a function to know if we are dealing with
an Intrinsic (like a PARTS one) or not.
Currently we ignore the results of all the Intrinsics.
This is important in the case of PARTS as we can introduce
redundant information or insert infinite number of
intrinsics with out current implementation.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>