PartsSpill: Fix bug when appending at end of MBB
At this point of the lowering, it does not holds that
we will have a terminator instruction at the end of basic
block.
We need to account that we may try to add the spill at the
end of machine basic block and we do not have machine
instruction as reference to insert before it.

This case may happen when the compiler is on the way of
merging to machine basic block, like in the case of our
getelementptr spill test.

( test/CodeGen/AArch64/AArch64Parts/AArch64PartsSpill/dpi-getelementptr-data-pointer.ll )

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>