PARTS:Spill:test: Add load/store optimizer test
Add a instrumented version of one of the AArch64 load/store
optimizer pass, to make certain that we do not break that
optimization with out intrinsics.

We should add more of these kind of tests later on for other
type of optimizations that are run before our final parts spill
ass.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>