RDF: Register Data Flow Analysis
Move the RDF files under the generic CodeGen directory.
We do this because we want to reuse that analysis for
AArch64 and our data pointer spilling support.

We will use the analysis to help us remove properly the
dead instructions (currently we can still leave dead
instructions around) and to know which callee save registers
need to be sign/auth before function calls.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>