PartsDpiForCSR: Extract a function
Extract a function to find out if the CSR holds a data pointer,
by looping through the loop of uses of the CSR and trying to
find our PARTS intrinsics.

Now we can also get rid off the awful break from the previous
commit, because now we have decoupled the find code from the
insertion code.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>