PARTS:Spill: Protect after CSR after local spill
Rework our logic to find our data pointer intrinsics.
Remove the ugly hack that check if the register spill
was the pointer and not the modifier in AUTDA, which
anyway was broken.
Now we have another Parts intrinsic to check for PARTS_RELOAD,
cause now a reloaded spilled maybe put in a CSR register.
So we need to do the things properly and split the logic
between finding definitions intrinsics and finding use intrinsics.
This also makes thing more clear and neat about what we are doing.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>