PARTS:Spill:test: Fix CSR protection tests.
We had a bug when trying to the AUTDA intrinsic.
Due to that bug, we were protecting less callee saved
registers that we should.
This fixes will account for that and avoid regressions
on such cases.
Also the freeing of the x23 register trigger some cosmetic
changes in the code. This commit unfortunately also accounts for
that, although the right thing to do is to split it in two.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>