Builtins:test: Check handling of PARTS modifier builtin
Users that one to manually PAC their pointers in a compatible
way with PARTS, needs access to the modifier associated to
the given pointer.
We will expose a new builtin function that allows the user
to get the modifier id for a given pointer.
This is the basic test for transforming our intrinsic call
to a 64 bit integer assignment.

Signed-off-by: Carlos Chinea <carlos.chinea.perez@huawei.com>